CRUD de pessoas completo

+++ src/main/java/br/com/tamarozzi/controller/PessoaController.java
@@ -9,6 +9,7 @@ import br.com.tamarozzi.model.Pessoa;
 import br.com.tamarozzi.service.impl.PessoaServiceImpl;
 import br.com.tamarozzi.services.PessoaService;
 import java.util.List;
+import javax.swing.JOptionPane;
 
 /**
  *
@@ -25,18 +26,27 @@ public class PessoaController {
     public List<Pessoa> getAllPessoa(String campo, String valor, String tipoPessoa, String situacao) {
         return this.pessoaService.getAllPessoa(campo, valor, tipoPessoa, situacao);
     }
-    
+
     public void editPessoa(Pessoa p) {
-        if(p.getId() > 0) {
+        if (p.getId() > 0) {
             this.pessoaService.edit(p);
-        } else {            
+        } else {
             this.pessoaService.add(p);
         }
     }
-    
+
     public void deletePessoa(List<Pessoa> pessoas) {
-        pessoas.stream().forEach((p) -> {
-            this.pessoaService.delete(p.getId());
-        });
+
+        int opcao = JOptionPane.showConfirmDialog(null,
+                "Confirma a exclusão deste(s) registro(s)?",
+                "Atenção",
+                JOptionPane.YES_NO_OPTION,
+                JOptionPane.INFORMATION_MESSAGE);
+
+        if (opcao == 0) {
+            pessoas.stream().forEach((p) -> {
+                this.pessoaService.delete(p.getId());
+            });
+        }
     }
 }
